+++ Lambda_Partition_Algorithm/partition_criterion_check.hpp
@@ -4,4 +4,4 @@
 
 using namespace std;
 
-optional<string> check_good_lambda (vector<uint32_t> lambda_partition, int N);
+optional<string> check_good_lambda (vector<uint64_t> lambda_partition, int N);
